move table into endpoint
