+++ backend/src/users/users.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UsersService } from './services/users.service';
 import { UsersController } from './controllers/users.controller';
-import { User } from 'src/Tables/User';
+import { User } from 'src/users/User';
 
 
 @Module({
